pinterest/requirements.txt
absl-py==1.1.0
tensorflow==2.9.1
typed-ast==1.5.4
typing_extensions==4.2.0
jax==0.3.15
flax==0.5.2
optax==0.1.2
